+++ CHANGELOG.md
@@ -6,9 +6,16 @@
 - Endpoint for updating custom tenant data changed (#98, PLUM Sprint 221104)
 - Unset credential phone/email by setting it to null instead of empty string (#117, PLUM Sprint 221104)
 
+### Fix
+- Logout with ID token (#116, PLUM Sprint 221104)
+- Disable registration service when no credential provider supports registration (#118, PLUM Sprint 221104)
+
 ### Features
 - Roles have an optional "description" field (#103, PLUM Sprint 221021)
 - User registration (invitation only) (#86, PLUM Sprint 221021)
+- Delete and rename resources (#113, PLUM Sprint 221104)
+- List roles that contain a specific resource (#113, PLUM Sprint 221104)
+- Include session ID and parent session ID in ID token (#116, PLUM Sprint 221104)
 
 ### Refactoring
 - Keep superuser role after provisioning (#102, PLUM Sprint 221021)